Fix type in rateLimit() signature - microtime() returns seconds as a float

+++ src/Origin/Discord.php
@@ -197,7 +197,7 @@ class Discord extends Origin
     }
 
     /** @see Https::get() for handling of 429s via `retry-after` header. */
-    private function rateLimit(array $headers, int $replySeconds = 0): void
+    private function rateLimit(array $headers, float $replySeconds = 0): void
     {
         // Get header info.
         $limit = $headers['x-ratelimit-limit'][0] ?? null; // total number that can be made
